MA_Nを正しく算出できるようにした

+++ app/services/technical/trend/ma_n.py
@@ -2,4 +2,5 @@ import pandas as pd
 
 def compute_ma_n(data: pd.DataFrame, n: int) -> float:
   close = data["Close"]
-  return close.rolling(window=n).mean().iloc[-1]
+  # TODO: floatがテストできていない実装になっている
+  return float(close.rolling(window=n).mean().iloc[-1])
